# Make HMAC authenticationCode and update inlinable