@baarde 2024-06-12
@Lukasa Thanks for your feedback. I'll use Feedback Assistant.

Would it be less ABI-impacting to make `update(bufferPointer:)` public?