@Lukasa 2024-06-17
Potentially, yes, it's worth suggesting in your Feedback Assistant report. Do you have the number handy?