@baarde 2024-06-26
@Lukasa I've submitted a feedback: FB14076813.